refactor(chat): relocate artifact URL helpers; tighten project gating
- Move isProjectArtifact + getArtifactApiUrl from
  StandaloneArtifactPreview.tsx to @/lib/api/artifacts (the
  presentational component is no longer the home of pure helpers).
  Three callers updated: ChatInputArea, ArtifactsPage, and the
  helpers test.
- ArtifactAttachmentCard project/session derivation now mirrors
  getArtifactApiUrl: only treat as project-scoped when projectId is
  actually set, so project artifacts missing a projectId fall back
  to a session-scoped fetch instead of producing an unfetchable
  request.
- StandaloneArtifactPreview hides Go-to-Project when projectId is
  missing (silent no-op replaced with no button). The card-level
  handlers drop their now-redundant if-checks.
- FileBadge: replace native <button> with the design-system Button
  (ghost variant) per repo conventions, padding stripped to keep
  the badge tight.
