+++ client/webui/frontend/src/lib/api/artifacts/service.ts
@@ -1,6 +1,28 @@
 import { api, getErrorFromResponse } from "@/lib/api";
 import { getArtifactContent as getArtifactContentUtil, validIdOrUndefined } from "@/lib/utils/file";
-import type { BulkArtifactsResponse } from "./types";
+import type { ArtifactWithSession, BulkArtifactsResponse } from "./types";
+
+/**
+ * True when the artifact belongs to a project rather than a chat session.
+ * Backend uses "project-{id}" format; "project:{id}" is kept for backward compat.
+ */
+export function isProjectArtifact(artifact: ArtifactWithSession): boolean {
+    return artifact.sessionId.startsWith("project:") || artifact.sessionId.startsWith("project-") || artifact.source === "project";
+}
+
+/**
+ * Resolve the correct API path for an artifact.
+ *
+ * For project artifacts we pass "null" as the session placeholder in the path
+ * and the actual project via the project_id query param — the backend endpoint
+ * requires a session_id path segment.
+ */
+export function getArtifactApiUrl(artifact: ArtifactWithSession): string {
+    if (isProjectArtifact(artifact) && artifact.projectId) {
+        return `/api/v1/artifacts/null/${encodeURIComponent(artifact.filename)}?project_id=${encodeURIComponent(artifact.projectId)}`;
+    }
+    return `/api/v1/artifacts/${encodeURIComponent(artifact.sessionId)}/${encodeURIComponent(artifact.filename)}`;
+}
 
 /**
  * Retrieves artifact content for preview, with ID validation and a "latest" version default.
